refactor!: improve database models
