+++ model/src/database/modlog.rs
@@ -1,6 +1,10 @@
 //! Models for the `modlogs` collection.
 
-use mongodb::bson::oid::ObjectId;
+use anyhow::anyhow;
+use mongodb::{
+    bson::{doc, oid::ObjectId, to_document, Bson},
+    Cursor,
+};
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, skip_serializing_none, DisplayFromStr};
 use time::OffsetDateTime;
@@ -12,11 +16,13 @@ use twilight_model::{
     util::ImageHash,
 };
 
+use super::DbClient;
 use crate::serde::{DateTimeAsBson, IdAsI64};
 
 /// Moderation log entry.
 ///
-/// This struct represent a unique moderation log entry stored in the database.
+/// This type represent a moderation log entry stored in the `modlogs` collection
+/// of the database.
 #[serde_as]
 #[skip_serializing_none]
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
@@ -68,4 +74,62 @@ pub struct ModlogUser {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde_as(as = "Option<DisplayFromStr>")]
     pub avatar: Option<ImageHash>,
+}
+
+// Implementation of methods to query the database.
+impl DbClient {
+    /// Insert a new [`Modlog`] in the database.
+    pub async fn create_modlog(&self, modlog: &Modlog) -> Result<ObjectId, anyhow::Error> {
+        let result = self
+            .db()
+            .collection::<Modlog>(Modlog::COLLECTION)
+            .insert_one(modlog, None)
+            .await?;
+
+        match result.inserted_id {
+            Bson::ObjectId(id) => Ok(id),
+            other => Err(anyhow!("expected object id, got {:?}", other)),
+        }
+    }
+
+    /// Get a [`Modlog`] from the database with its id.
+    pub async fn get_modlog(&self, id: ObjectId) -> Result<Option<Modlog>, anyhow::Error> {
+        let query = doc! { "_id": id };
+
+        let modlog = self
+            .db()
+            .collection::<Modlog>(Modlog::COLLECTION)
+            .find_one(query, None)
+            .await?;
+
+        Ok(modlog)
+    }
+
+    /// Find multiple [`Modlog`]s from the database that match a given guild id
+    /// and optional user id.
+    pub async fn find_modlogs(
+        &self,
+        guild_id: Id<GuildMarker>,
+        user_id: Option<Id<UserMarker>>,
+    ) -> Result<Cursor<Modlog>, anyhow::Error> {
+        let query = ModlogQuery { guild_id, user_id };
+
+        let cursor = self
+            .db()
+            .collection::<Modlog>(Modlog::COLLECTION)
+            .find(to_document(&query)?, None)
+            .await?;
+
+        Ok(cursor)
+    }
+}
+
+/// Query modlogs with guild_id and optional user_id
+#[serde_as]
+#[derive(Debug, Serialize, Deserialize, Clone, Copy)]
+struct ModlogQuery {
+    #[serde_as(as = "IdAsI64")]
+    pub guild_id: Id<GuildMarker>,
+    #[serde_as(as = "Option<IdAsI64>")]
+    pub user_id: Option<Id<UserMarker>>,
 }
